requirements.txt
#tensorflow==2.15.0
tensorflow
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras